W8: Refactored DR variables into their own files in  &
