--- test/dr/test_diabetic_retinopathy_pattern.py
@@ -1,4 +1,0 @@
-import pytest
-
-
-# TODO: Create tests for regex patterns
